+++ packages/app-utils/package.json
@@ -1,0 +1,77 @@
+{
+  "name": "@red-hat-developer-hub/app-utils",
+  "version": "1.0.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.esm.js",
+    "module": "dist/index.esm.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "common-library",
+    "supported-versions": "1.0.0",
+    "pluginId": "app-utils",
+    "pluginPackage": "@red-hat-developer-hub/app-utils",
+    "pluginPackages": [
+      "@red-hat-developer-hub/app-utils"
+    ]
+  },
+  "scripts": {
+    "build": "backstage-cli package build",
+    "clean": "backstage-cli package clean",
+    "test": "backstage-cli package test --passWithNoTests --coverage",
+    "lint:check": "backstage-cli package lint",
+    "lint:fix": "backstage-cli package lint --fix",
+    "tsc": "tsc",
+    "prettier:check": "prettier --ignore-unknown --check .",
+    "prettier:fix": "prettier --ignore-unknown --write ."
+  },
+  "dependencies": {
+    "@backstage/catalog-model": "1.7.5",
+    "@backstage/config": "1.3.3",
+    "@backstage/core-app-api": "1.18.0",
+    "@backstage/core-plugin-api": "1.10.9",
+    "@backstage/plugin-api-docs": "0.12.10",
+    "@backstage/plugin-catalog": "1.31.2",
+    "@backstage/plugin-catalog-import": "0.13.4",
+    "@backstage/plugin-org": "0.6.43",
+    "@backstage/plugin-scaffolder": "1.34.0",
+    "@mui/material": "5.18.0",
+    "lodash": "4.17.21"
+  },
+  "devDependencies": {
+    "@backstage/cli": "0.34.1",
+    "@backstage/test-utils": "1.7.11",
+    "@testing-library/dom": "9.3.4",
+    "@testing-library/jest-dom": "6.9.1",
+    "@testing-library/react": "14.3.1",
+    "@testing-library/react-hooks": "8.0.1",
+    "@types/node": "22.19.1",
+    "@types/react": "18.3.27",
+    "@types/react-dom": "18.3.7",
+    "prettier": "3.7.4",
+    "react": "18.3.1",
+    "typescript": "5.9.3"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "files": [
+    "dist"
+  ],
+  "peerDependencies": {
+    "react": "16.13.1 || ^17.0.0 || ^18.2.0"
+  }
+}
