+++ packages/app-utils/src/dynamic/config/extractDynamicConfig.ts
@@ -1,219 +1,29 @@
-import { Entity } from '@backstage/catalog-model';
-import { ApiHolder } from '@backstage/core-plugin-api';
-import { isKind } from '@backstage/plugin-catalog';
-
-import {
-  MountPointConfigRaw,
-  MountPointConfigRawIf,
-  RouteBinding,
-} from '@red-hat-developer-hub/plugin-utils';
-
-import { hasAnnotation, isType } from '../../components/catalog/utils';
-import { DynamicTranslationResource } from '../../types/types';
 import { extractMenuItems } from './extractDynamicConfigFrontend';
-
-export type DynamicRouteMenuItem =
-  | {
-      text: string;
-      textKey?: string;
-      icon: string;
-      parent?: string;
-      priority?: number;
-      enabled?: boolean;
-    }
-  | {
-      module?: string;
-      importName: string;
-      config?: {
-        props?: Record<string, any>;
-      };
-    };
-
-export type MenuItemConfig = {
-  icon?: string;
-  title?: string;
-  priority?: number;
-  parent?: string;
-};
-
-export type MenuItem = {
-  name: string;
-  title: string;
-  titleKey?: string;
-  icon: string;
-  children: MenuItem[];
-  priority?: number;
-  to?: string;
-  parent?: string;
-  enabled?: boolean;
-};
-
-export type DynamicRoute = {
-  scope: string;
-  module: string;
-  importName: string;
-  path: string;
-  menuItem?: DynamicRouteMenuItem;
-  config?: {
-    props?: Record<string, any>;
-  };
-};
-
-type PluginModule = {
-  scope: string;
-  module: string;
-};
-
-type MountPoint = {
-  scope: string;
-  mountPoint: string;
-  module: string;
-  importName: string;
-  config?: MountPointConfigRaw;
-};
-
-type AppIcon = {
-  scope: string;
-  name: string;
-  module: string;
-  importName: string;
-};
-
-type BindingTarget = {
-  scope: string;
-  name: string;
-  module: string;
-  importName: string;
-};
-
-type ApiFactory = {
-  scope: string;
-  module: string;
-  importName: string;
-};
-
-type AnalyticsApiExtension = {
-  scope: string;
-  module: string;
-  importName: string;
-};
-
-type ScaffolderFieldExtension = {
-  scope: string;
-  module: string;
-  importName: string;
-};
-
-type TechdocsAddon = {
-  scope: string;
-  module: string;
-  importName: string;
-  config?: {
-    props?: Record<string, any>;
-  };
-};
-
-type EntityTab = {
-  mountPoint: string;
-  path: string;
-  title: string;
-  titleKey?: string;
-  pariority?: number;
-};
-
-type EntityTabEntry = {
-  scope: string;
-  mountPoint: string;
-  path: string;
-  title: string;
-  titleKey?: string;
-  priority?: number;
-};
-
-type ThemeEntry = {
-  scope: string;
-  module: string;
-  id: string;
-  title: string;
-  variant: 'light' | 'dark';
-  icon: string;
-  importName: string;
-};
-
-type SignInPageEntry = {
-  scope: string;
-  module: string;
-  importName: string;
-};
-
-type ProviderSetting = {
-  title: string;
-  description: string;
-  provider: string;
-};
-
-type CustomProperties = {
-  pluginModule?: string;
-  dynamicRoutes?: {
-    importName?: string;
-    module?: string;
-    scope?: string;
-    path: string;
-    menuItem?: DynamicRouteMenuItem;
-  }[];
-  menuItems?: { [key: string]: MenuItemConfig };
-  routeBindings?: {
-    targets: BindingTarget[];
-    bindings: RouteBinding[];
-  };
-  entityTabs?: EntityTab[];
-  mountPoints?: MountPoint[];
-  appIcons?: AppIcon[];
-  apiFactories?: ApiFactory[];
-  analyticsApiExtensions?: AnalyticsApiExtension[];
-  providerSettings?: ProviderSetting[];
-  scaffolderFieldExtensions?: ScaffolderFieldExtension[];
-  signInPage: SignInPageEntry;
-  techdocsAddons?: TechdocsAddon[];
-  themes?: ThemeEntry[];
-  translationResources?: DynamicTranslationResource[];
-};
-
-export type FrontendConfig = {
-  [key: string]: CustomProperties;
-};
-
-export type DynamicPluginConfig = {
-  frontend?: FrontendConfig;
-};
-
-type DynamicConfig = {
-  pluginModules: PluginModule[];
-  apiFactories: ApiFactory[];
-  analyticsApiExtensions: AnalyticsApiExtension[];
-  appIcons: AppIcon[];
-  dynamicRoutes: DynamicRoute[];
-  menuItems: MenuItem[];
-  entityTabs: EntityTabEntry[];
-  mountPoints: MountPoint[];
-  providerSettings: ProviderSetting[];
-  routeBindings: RouteBinding[];
-  routeBindingTargets: BindingTarget[];
-  scaffolderFieldExtensions: ScaffolderFieldExtension[];
-  signInPages: SignInPageEntry[];
-  techdocsAddons: TechdocsAddon[];
-  themes: ThemeEntry[];
-  translationResources: DynamicTranslationResource[];
-};
+import {
+  AnalyticsApiExtension,
+  ApiFactory,
+  AppIcon,
+  BindingTarget,
+  DynamicConfig,
+  DynamicModuleEntry,
+  DynamicPluginConfig,
+  DynamicRoute,
+  DynamicTranslationResource,
+  EntityTabEntry,
+  MountPoint,
+  ProviderSetting,
+  RouteBinding,
+  ScaffolderFieldExtension,
+  SignInPageEntry,
+  TechdocsAddon,
+  ThemeEntry,
+} from './types';
 
 /**
  * Converts the dynamic plugin configuration structure to the data structure
  * required by the dynamic UI, substituting in any defaults as needed
  */
-function extractDynamicConfig(
-  dynamicPlugins: DynamicPluginConfig = { frontend: {} },
-) {
-  const frontend = dynamicPlugins.frontend || {};
+export function extractDynamicConfig(dynamicPlugins: DynamicPluginConfig) {
   const config: DynamicConfig = {
     pluginModules: [],
     apiFactories: [],
@@ -232,6 +42,10 @@ function extractDynamicConfig(
     themes: [],
     translationResources: [],
   };
+  if (dynamicPlugins?.frontend === undefined) {
+    return config;
+  }
+  const { frontend } = dynamicPlugins;
   config.signInPages = Object.entries(frontend).reduce<SignInPageEntry[]>(
     (pluginSet, [scope, { signInPage }]) => {
       if (!signInPage) {
@@ -250,7 +64,7 @@ function extractDynamicConfig(
     },
     [],
   );
-  config.pluginModules = Object.entries(frontend).reduce<PluginModule[]>(
+  config.pluginModules = Object.entries(frontend).reduce<DynamicModuleEntry[]>(
     (pluginSet, [scope, customProperties]) => {
       pluginSet.push({
         scope,
@@ -425,57 +239,3 @@ function extractDynamicConfig(
 
   return config;
 }
-
-/**
- * Evaluate the supplied conditional map.  Used to determine the visibility of
- * tabs in the UI
- * @param conditional
- * @returns
- */
-export function configIfToCallable(conditional: MountPointConfigRawIf) {
-  return (entity: Entity, context?: { apis: ApiHolder }) => {
-    if (conditional?.allOf) {
-      return conditional.allOf
-        .map(conditionsArrayMapper)
-        .every(f => f(entity, context));
-    }
-    if (conditional?.anyOf) {
-      return conditional.anyOf
-        .map(conditionsArrayMapper)
-        .some(f => f(entity, context));
-    }
-    if (conditional?.oneOf) {
-      return (
-        conditional.oneOf
-          .map(conditionsArrayMapper)
-          .filter(f => f(entity, context)).length === 1
-      );
-    }
-    return true;
-  };
-}
-
-export function conditionsArrayMapper(
-  condition:
-    | {
-        [key: string]: string | string[];
-      }
-    | Function,
-): (entity: Entity, context?: { apis: ApiHolder }) => boolean {
-  if (typeof condition === 'function') {
-    return (entity: Entity, context?: { apis: ApiHolder }): boolean =>
-      condition(entity, context);
-  }
-  if (condition.isKind) {
-    return isKind(condition.isKind);
-  }
-  if (condition.isType) {
-    return isType(condition.isType);
-  }
-  if (condition.hasAnnotation) {
-    return hasAnnotation(condition.hasAnnotation as string);
-  }
-  return () => false;
-}
-
-export default extractDynamicConfig;
