+++ packages/app-utils/turbo.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["//"],
+  "tasks": {
+    "tsc": { "outputs": ["../../dist-types/packages/app-utils/**"] }
+  }
+}
