+++ packages/app/package.json
@@ -46,6 +46,7 @@
     "@mui/icons-material": "5.18.0",
     "@mui/material": "5.18.0",
     "@mui/styled-engine": "5.18.0",
+    "@red-hat-developer-hub/app-utils": "workspace:*",
     "@red-hat-developer-hub/backstage-plugin-theme": "0.10.2",
     "@red-hat-developer-hub/backstage-plugin-translations": "0.0.2",
     "@red-hat-developer-hub/plugin-utils": "1.0.0",
