+++ packages/app/src/components/DynamicRoot/ScalprumRoot.tsx
@@ -6,14 +6,19 @@ import { AppConfig } from '@backstage/config';
 import { ConfigReader, defaultConfigLoader } from '@backstage/core-app-api';
 import { AnyApiFactory } from '@backstage/core-plugin-api';
 
-import { DynamicRootConfig } from '@red-hat-developer-hub/plugin-utils';
+import {
+  DynamicPluginConfig,
+  overrideBaseUrlConfigs,
+  TranslationConfig,
+} from '@red-hat-developer-hub/app-utils';
+import {
+  DynamicRootConfig,
+  StaticPlugins,
+} from '@red-hat-developer-hub/plugin-utils';
 import { AppsConfig } from '@scalprum/core';
 import { ScalprumProvider } from '@scalprum/react-core';
 
-import { TranslationConfig } from '../../types/types';
-import { DynamicPluginConfig } from '../../utils/dynamicUI/extractDynamicConfig';
-import overrideBaseUrlConfigs from '../../utils/dynamicUI/overrideBaseUrlConfigs';
-import { DynamicRoot, StaticPlugins } from './DynamicRoot';
+import { DynamicRoot } from './DynamicRoot';
 import Loader from './Loader';
 
 export type ScalprumApiHolder = {
