+++ packages/app/src/components/Root/Root.tsx
@@ -38,6 +38,7 @@ import Collapse from '@mui/material/Collapse';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import { styled, SxProps, Theme } from '@mui/material/styles';
+import { MenuIcon } from '@red-hat-developer-hub/app-utils';
 import { ThemeConfig } from '@red-hat-developer-hub/backstage-plugin-theme';
 import DynamicRootContext, {
   ResolvedMenuItem,
@@ -47,7 +48,6 @@ import { useLanguagePreference } from '../../hooks/useLanguagePreference';
 import { useTranslation } from '../../hooks/useTranslation';
 import { ApplicationHeaders } from './ApplicationHeaders';
 import { CustomSidebarItem } from './CustomSidebarItem';
-import { MenuIcon } from './MenuIcon';
 import { SidebarLogo } from './SidebarLogo';
 
 /**
