+++ packages/app/src/components/catalog/EntityPage/ContextMenuAwareEntityLayout.tsx
@@ -8,8 +8,9 @@ import {
 
 import { EntityLayout } from '@backstage/plugin-catalog';
 
+import { MenuIcon } from '@red-hat-developer-hub/app-utils';
+
 import getMountPointData from '../../../utils/dynamicUI/getMountPointData';
-import { MenuIcon } from '../../Root/MenuIcon';
 
 const makeIcon = (iconName: string) => () => <MenuIcon icon={iconName} />;
