+++ packages/app/src/types/types.ts
@@ -1,5 +1,3 @@
-import { TranslationResource } from '@backstage/core-plugin-api/alpha';
-
 export type LearningPathLink = {
   label: string;
   url: string;
@@ -16,39 +14,3 @@ export type BuildInfo = {
   full?: boolean;
   overrideBuildInfo?: boolean;
 };
-
-export type TranslationConfig = {
-  defaultLocale?: string;
-  locales: string[];
-  overrides?: string[];
-};
-
-export type JSONTranslationConfig = {
-  locale: string;
-  path: string;
-};
-
-export type DynamicTranslationResource = {
-  scope: string;
-  module: string;
-  importName: string;
-  ref?: string | null;
-  jsonTranslations?: JSONTranslationConfig[];
-};
-
-// Types from Backstage core-plugin-api do not expose loader function type
-// so we need to create our own internal types to access the loader function
-
-type InternalTranslationResourceLoader = () => Promise<{
-  messages: { [key in string]: string | null };
-}>;
-
-export interface InternalTranslationResource<
-  TId extends string = string,
-> extends TranslationResource<TId> {
-  version: 'v1';
-  resources: {
-    language: string;
-    loader: InternalTranslationResourceLoader;
-  }[];
-}
