+++ packages/app/src/utils/translations/translationResourceProcessor.ts
@@ -3,7 +3,8 @@ import {
   TranslationResource,
 } from '@backstage/core-plugin-api/alpha';
 
-import { InternalTranslationResource } from '../../types/types';
+import { InternalTranslationResource } from '@red-hat-developer-hub/app-utils';
+
 import { translationResourceGenerator } from './translationResourceGenerator';
 
 export interface TranslationResourceWithRef {
